Add file structure and code cleanup
