--- src/Form.js
@@ -1,15 +1,0 @@
-import React, { useState } from 'react'
-
-export default function Form({ onSearchSubmit }){
-  const [searchValue, setSearch] = useState(null)
-  console.log(searchValue)
-  return (
-    <div>
-        <label>
-          GIF:<br />
-          <input type="text" onChange={e => setSearch(e.target.value)}/>
-        </label>
-        <button onClick={() => onSearchSubmit(searchValue)}>Search!</button>
-    </div>
-  )
-}
